working JBEAP_133375 client

+++ scripts/stopConfiguration.sh
@@ -9,11 +9,12 @@ while [ $# -gt 0 ]; do
   shift
 done
 
+echo
+
 NAME=$c
-echo "New configuration name: $NAME"
+echo "*** $NAME ***"
 
 TMP_PREFIX=$tp
-echo "Used tmp prefix: $TMP_PREFIX"
 
 #$EAP_HOME/bin/$NAME.sh -c $PROFILE-for-$NAME.xml -b 127.0.0.1 -Djboss.server.base.dir=$EAP_HOME-$NAME
 
@@ -23,18 +24,22 @@ file="/tmp/${TMP_PREFIX}_running_${NAME}.txt"
 #if process is running, exit
 if [ -f "$file" ]
 then
-	echo "killing ..."
+
+	parent_process=`cat $file`
+    #parent_process=$value
+
+    if [ $parent_process == 'cmd' ]
+    then
+        echo "!!!!!!!!!!!!!! started from cmd !!!!!!!!!!!!!!"
+    else
+        echo "killing ..."
+        pkill -9 -P $parent_process
+        kill -9 $parent_process
+        rm $file
+        echo STOPPED
+    fi
 else
-	echo "$file not found."
-	echo "CAN NOT STOP."
+	echo "$file not found. already stopped"
 	exit 1
 fi
 
-parent_process=`cat $file`
-#parent_process=$value
-
-pkill -9 -P $parent_process
-kill -9 $parent_process
-
-rm $file
-echo STOPPED
